update: reorganized bytecode tool util folder
